CADC-9463 (#44)
* CADC-9337 - add store_newer_files_only flag to manage_composable.Config.

* CADC-9337 - execute_composable.CaomExecute._cadc_put will check timestamps before storing, if configured to do so.

* CADC-9337 - update module version.

* CADC-9337 - improve logging message content.

* CADC-9337 - ensure get_file_info call has the correct parameters.

* CADC-9337 - repair the date handling for the format returned from CadcDataClient.get_file_info.

* CADC-9337 - add a log message to indicate transfer was considered.

* CADC-9337 - address code review comments.

* GemProc test run - there's a better way to capture failure messages.

* CADC-9337 - better failure logging handling, because I understand more about python.

* CADC-9337 - CFHT edge cases.

* CADC-8551 - yes, I know, wrong reference on the wrong branch.

* CADC-9463 - add a method to build up chunk-level TemporalWCS instances with a CoordBounds1D member.

* CADC-9287 - address code review comments.

* CADC-9287 - add some handling for repairing, and ignoring, a value that is originally None.

* CADC-9463 - clean up spurious logging.error messages.

* CADC-9463 - use plane.provenance.inputs to find TemporalWCS samples, instead of header keywords.

* CADC-9463 - compare to the enumerated type, not the string for part.product_type.

* CADC-9463 - add class VoFitsTransfer to use FITS checking with VoTransfer.

* CADC-9463 - add VaultListDirDataSource to data_source_composable.

* CADC-9463 - a Transfer implementation that will not have the _cadc_client data member assigned the default value.

* CADC-9463 - make sure check gets executed.

* CADC-9463 - test coverage, and checking, for VoFitsTransfer.

* CADC-9463 - add test coverage for VaultListDirDataSource.

+++ caom2pipe/transfer_composable.py
@@ -70,11 +70,11 @@
 import logging
 import os
 
-from cadcdata import CadcDataClient
 from caom2pipe import manage_composable as mc
 
 
-__all__ = ['CadcTransfer', 'Transfer', 'VoTransfer']
+__all__ = ['CadcTransfer', 'FtpTransfer', 'HttpTransfer', 'Transfer',
+           'VoFitsTransfer', 'VoTransfer']
 
 
 class Transfer(object):
@@ -177,6 +177,14 @@ class FitsTransfer(Transfer):
         self._observable = None
         self._logger = logging.getLogger(self.__class__.__name__)
 
+    @property
+    def observable(self):
+        return self._observable
+
+    @observable.setter
+    def observable(self, value):
+        self._observable = value
+
     def get(self, source, dest_fqn):
         raise NotImplementedError
 
@@ -189,21 +197,25 @@ class FitsTransfer(Transfer):
                 h.verify('warn')
             hdulist.close()
         except (fits.VerifyError, OSError) as e:
-            self._observable.rejected.record(mc.Rejected.BAD_DATA,
-                                             os.path.basename(dest_fqn))
-            os.unlink(dest_fqn)
-            raise mc.CadcException(
-                f'astropy verify error {dest_fqn} when reading {e}')
+            if self._observable is not None:
+                self._observable.rejected.record(mc.Rejected.BAD_DATA,
+                                                 os.path.basename(dest_fqn))
+            if os.path.exists(dest_fqn):
+                os.unlink(dest_fqn)
+                raise mc.CadcException(
+                    f'astropy verify error {dest_fqn} when reading {e}')
         # a second check that fails for some NEOSSat cases - if this works,
         # the file might have been correctly retrieved
         try:
             # ignore the return value - if the file is corrupted, the getdata
             # fails, which is the only interesting behaviour here
             fits.getdata(dest_fqn, ext=0)
-        except TypeError as e:
-            self._observable.rejected.record(mc.Rejected.BAD_DATA,
-                                             os.path.basename(dest_fqn))
-            os.unlink(dest_fqn)
+        except (TypeError, OSError) as e:
+            if self._observable is not None:
+                self._observable.rejected.record(mc.Rejected.BAD_DATA,
+                                                 os.path.basename(dest_fqn))
+            if os.path.exists(dest_fqn):
+                os.unlink(dest_fqn)
             raise mc.CadcException(
                 f'astropy getdata error {dest_fqn} when reading {e}')
 
@@ -248,6 +260,24 @@ class FtpTransfer(FitsTransfer):
         """
         self._logger.debug(f'Retrieve {source}')
         mc.ftp_get_timeout(self._ftp_host, source, dest_fqn)
+        if '.fits' in dest_fqn:
+            self.check(dest_fqn)
+        self._logger.debug(f'Successfully retrieved {source}')
+
+
+class VoFitsTransfer(FitsTransfer):
+    """
+    Uses the vos Client to manage transfers from CADC to local disk. Have
+    FITS integrity-checking.
+    """
+
+    def __init__(self, vos_client):
+        super(VoFitsTransfer, self).__init__()
+        self._vos_client = vos_client
+        self._logger = logging.getLogger(self.__class__.__name__)
+
+    def get(self, source, dest_fqn):
+        self._vos_client.copy(source, dest_fqn, send_md5=True)
         if '.fits' in dest_fqn:
             self.check(dest_fqn)
         self._logger.debug(f'Successfully retrieved {source}')
